requirements.txt
streamlit
Pillow
matplotlib
numpy
pandas